Refactor: rename composable and component by type of using
