+++ src/features/Play/Quize/components/Quize.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
   import { DictionaryItem } from '@/core/models/Topic';
-  import { QuestionItem } from '../composables/useQuize';
+  import { QuestionItem } from '../composables/useWordVariants';
   import { Colors } from '@/core/models/enums';
 
   import { computed } from 'vue';
