+++ src/features/Play/Quize/components/QuizeWordVariants.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-  import { QuestionItem } from '../composables/useQuize';
+  import { QuestionItem } from '../composables/useWordVariants';
   import { Colors, Variants } from '@/core/models/enums';
 
   defineProps<{
